CreateServlet
Wrote the first half of create servlet where the user can add a
question to the question list.  About to finish the second half.  I
noticed that some questions had double as pointValue and some had int,
so I switched those to doubles.

+++ src/model/MultipleChoice.java
@@ -7,11 +7,11 @@ public class MultipleChoice extends Question {
 
 	private String question;
 	private String[] choices;
-	private int pointValue;
+	private double pointValue;
 	private Answers answer; 
 	
 	//Multiple choice questions support one question and one correct answer among a string of choices
-	public MultipleChoice(String question, String[] choices, ArrayList<Set<String>> answer, int pointValue){
+	public MultipleChoice(String question, String[] choices, ArrayList<Set<String>> answer, double pointValue){
 		this.question = question; 
 		this.answer = new Answers(answer);
 		this.pointValue = pointValue;
